requirements.txt
cftime==1.6.2
netCDF4==1.6.2
numpy==1.23.4
Pillow==9.3.0
